Refactoring file location
