+++ src/components/DetailCard.tsx
